+++ js/plugins/validator.js
@@ -370,10 +370,9 @@ var Validator = {
             val: 0,
             log: []
         };
-        var data = {};
+        var formData = Utils.formData(element);
 
         $.each(inputs, function(){
-            if (Utils.isValue(this.name)) data[this.name] = this.value;
             ValidatorFuncs.validate(this, result, o.onValidate, o.onError, o.requiredMode);
         });
 
@@ -381,16 +380,16 @@ var Validator = {
 
         element[0].action = this._action;
 
-        result.val += Utils.exec(o.onBeforeSubmit, [element], this.elem) === false ? 1 : 0;
+        result.val += Utils.exec(o.onBeforeSubmit, [element, formData], this.elem) === false ? 1 : 0;
 
         if (result.val === 0) {
-            Utils.exec(o.onValidateForm, [element, data], form);
+            Utils.exec(o.onValidateForm, [element, formData], form);
             setTimeout(function(){
-                Utils.exec(o.onSubmit, [element], form);
+                Utils.exec(o.onSubmit, [element, formData], form);
                 if (that._onsubmit !==  null) Utils.exec(that._onsubmit, null, form);
             }, o.submitTimeout);
         } else {
-            Utils.exec(o.onErrorForm, [result.log, element, data], form);
+            Utils.exec(o.onErrorForm, [result.log, element, formData], form);
             if (o.clearInvalid > 0) {
                 setTimeout(function(){
                     $.each(inputs, function(){