+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 4.2.28
 + [x] Docs: fix mistakes in table options
 + [x] Utils css: fix class `m4-cloak` 
-+ [x] Validator: add argument `data` to events `data-on-validate-form`, `data-on-error-form`. Data is a `object` and contains pairs: `input-name: input-value`  
++ [x] Validator: add argument `data` to events `data-on-validate-form`, `data-on-error-form`, `data-on-submit`. Data is a `object` and contains pairs: `input-name: input-value` for form elements. 
 
 ### 4.2.27
 + [x] Input: remove -webkit-autofill background color